Initial basic pass template

+++ tdc-api/src/main/resources/db/migration/V20211206A__page_template.sql
@@ -1,5 +1,5 @@
 create table if not exists `tdc_page_template` (
-    `id` int not null primary key auto_increment,
+    `id` bigint not null primary key auto_increment,
     `code` varchar(255) not null unique,
     `content` LONGTEXT
 );
